requirements.txt
python package
    biopython = 1.78
    pandas = 1.4.2
    argparse = 1.1
    os 

software
    fastqc = 0.11.9
    trim_galore = 0.6.7
